+++ state_machine/thread_test/Main.cpp
@@ -4,30 +4,8 @@
 #include<mutex>
 #include<chrono>
 #include<unordered_map>
+#include"SensorsRead.h"
 
-
-class Sensors
-{
-	int sensor_value;
-	std::mutex mutex;
-public:
-	Sensors() :sensor_value(0){}
-	int get_value()   { 	return sensor_value; }
-	void add_value(int money)
-	{
-		std::lock_guard<std::mutex> lockGuard(mutex);
-		for(int i = 0; i < money; ++i)
-		{	
-			sensor_value++;
-			std::this_thread::sleep_for (std::chrono::seconds(1));
-			if (i == 100) {
-				std::cout << "end";
-				std::terminate();
-			}
-		}
-	}
-
-};
